requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
boto3>=1.28.0
botocore>=1.31.0
psycopg2-binary>=2.9.0
